+++ doi-transformer/src/main/java/no/unit/nva/doi/transformer/DataciteResponseConverter.java
@@ -16,8 +16,8 @@ import java.util.UUID;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
-import no.unit.nva.doi.fetch.language.LanguageDetector;
-import no.unit.nva.doi.fetch.language.SimpleLanguageDetector;
+import no.unit.nva.doi.transformer.language.LanguageDetector;
+import no.unit.nva.doi.transformer.language.SimpleLanguageDetector;
 import no.unit.nva.doi.transformer.model.datacitemodel.DataciteContainer;
 import no.unit.nva.doi.transformer.model.datacitemodel.DataciteCreator;
 import no.unit.nva.doi.transformer.model.datacitemodel.DataciteRelatedIdentifier;
