+++ doi-transformer/src/main/java/no/unit/nva/doi/transformer/language/LanguageDetector.java
@@ -1,10 +1,10 @@
-package no.unit.nva.doi.fetch.language;
+package no.unit.nva.doi.transformer.language;
 
 import java.net.URI;
 import java.util.Locale;
 import java.util.Optional;
 
-import no.unit.nva.doi.fetch.language.exceptions.LanguageUriNotFoundException;
+import no.unit.nva.doi.transformer.language.exceptions.LanguageUriNotFoundException;
 import nva.commons.core.JacocoGenerated;
 
 public interface LanguageDetector {
