+++ doi-transformer/src/main/java/no/unit/nva/doi/transformer/language/exceptions/LanguageUriNotFoundException.java
@@ -1,4 +1,4 @@
-package no.unit.nva.doi.fetch.language.exceptions;
+package no.unit.nva.doi.transformer.language.exceptions;
 
 
 public class LanguageUriNotFoundException extends Exception {
