+++ doi-transformer/src/test/java/no/unit/nva/doi/transformer/CrossRefConverterTest.java
@@ -30,8 +30,8 @@ import java.util.Objects;
 import java.util.UUID;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import no.unit.nva.doi.fetch.language.LanguageMapper;
-import no.unit.nva.doi.fetch.language.exceptions.LanguageUriNotFoundException;
+import no.unit.nva.doi.transformer.language.LanguageMapper;
+import no.unit.nva.doi.transformer.language.exceptions.LanguageUriNotFoundException;
 import no.unit.nva.doi.transformer.model.crossrefmodel.CrossRefDocument;
 import no.unit.nva.doi.transformer.model.crossrefmodel.CrossrefAffiliation;
 import no.unit.nva.doi.transformer.model.crossrefmodel.CrossrefApiResponse;
